feat: integrate sanctuary grand milestone rewards
